% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baBatches}
\alias{baBatches}
\title{Get batch sizes to get the total n}
\usage{
baBatches(n, bsize)
}
\arguments{
\item{n}{sample size}

\item{bsize}{batch size}
}
\description{
Get batch sizes to get the total n
}
