% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bacProb}
\alias{bacProb}
\title{Probability of (n = n, r = 0:n)}
\usage{
bacProb(n, bsize, pmat)
}
\arguments{
\item{n}{sample size}

\item{bsize}{batch size}

\item{pmat}{probability matrix}
}
\description{
Probability of (n = n, r = 0:n)
}
