% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bacSimonDesign}
\alias{bacSimonDesign}
\title{Simon's two-stage design}
\usage{
bacSimonDesign(y0, y1, nmax, nmin, bsize, alpha, beta)
}
\arguments{
\item{y0}{response 0}

\item{y1}{response 1}

\item{nmax}{maximum size}

\item{nmin}{minimum size}

\item{bsize}{batch size}

\item{alpha}{observed alpha from simulations}

\item{beta}{observed beta from simulations.}
}
\description{
Simon's two-stage design
}
