% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bacSimonSingle}
\alias{bacSimonSingle}
\title{Single Simon 2-stage}
\usage{
bacSimonSingle(cumu, n1, r1, n, r)
}
\description{
@param cumu cumulative binomial distribution
 @param n1 first stage sample size
 @param r1 first stage response rate
 @param n total sample size
 @param r  second stage response rate
}
