% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_design.R
\name{baDesignSens}
\alias{baDesignSens}
\title{Evaluate design sensitivity}
\usage{
baDesignSens(rhos, design, ..., seed = NULL)
}
\arguments{
\item{design}{A list with p0, p1, batch size, n1, r1, n, r}
}
\description{
Get type I error and power given p0, p1, rho, batch size, n1, r1, n, r
}
