% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_design.R
\name{baDesignSensSingle}
\alias{baDesignSensSingle}
\title{Evaluate design sensitivity}
\usage{
baDesignSensSingle(rho, p0, p1, bsize, n1, r1, nt, r, ...)
}
\description{
Get type I error and power given p0, p1, rho, batch size, n1, r1, n, r for
one rho
}
