% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_tools.R
\name{baElicitFitMix}
\alias{baElicitFitMix}
\title{Fit mixed effect model to evaluate batch effect}
\usage{
baElicitFitMix(dat, fml, des_mu = 1, nbs = 1000)
}
\arguments{
\item{dat}{existing experiment data}

\item{fml}{mixed model formula that has batch factor with random effect}

\item{des_mu}{design vector for obtaining mean. By default, this will be the intercept}

\item{nbs}{number of bootstraps}
}
\value{
A matrix with 3 columns and \code{nbs} rows. The columns correspond
    to 1) standard deviation of the batch effect, 2) standard deviation of
    the random error, and 3) mean
}
\description{
Fit mixed effect model to evaluate batch effect
}
