% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_tools.R
\name{baElicitICC}
\alias{baElicitICC}
\title{Elicit ICC based on batch effect and random error standard deviation}
\usage{
baElicitICC(
  pars,
  p_resp,
  bsize = 5,
  beta = NULL,
  ntest = 10000,
  threshold = 2,
  take_exp = TRUE,
  seed = NULL
)
}
\arguments{
\item{pars}{parameters with 3 columns. 1) std_batch std of the batch effect
2) std_err: std of the random error and 3) mu intercept in the
proliferation model}

\item{p_resp}{response rate}

\item{bsize}{batch size}

\item{beta}{beta coefficient in the proliferation model}

\item{ntest}{number of replications}

\item{threshold}{threshold to be considered response}

\item{take_exp}{whether the proliferation model is log transformed}
}
\value{
Estimated ICC with its (bootstrap) SD.
}
\description{
For vaccine studies using pre- vs. post_vaccine proliferation ratio as the
endpoint
}
