% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baGetBvrCvIcc}
\alias{baGetBvrCvIcc}
\title{Get CV, batch effect variance ratio and ICC}
\usage{
baGetBvrCvIcc(par.err, par.other, rates, f.simu = baSimuTcell, ...)
}
\description{
Get CV, batch effect variance ratio and ICC
}
