% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baGetCuts}
\alias{baGetCuts}
\title{Get cut off of the outcome to get given response rates}
\usage{
baGetCuts(par.err, par.other, rates, f.simu = baSimuTcell, ..., seed = NULL)
}
\arguments{
\item{par.err}{a list of gamma, delta, epsilon error}

\item{par.other}{other parameters including u0, u1, v, and beta}

\item{rates}{response rates}

\item{f.simu}{simulating function}

\item{...}{reserved parameters}

\item{seed}{Seed}
}
\value{
a list of cut off given the response rates
}
\description{
Get cut off of the outcome to get given response rates
}
