% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_tools.R
\name{baGetOutcome}
\alias{baGetOutcome}
\title{Define outcome based on longitudinal outcomes}
\usage{
baGetOutcome(mat.y, type = c("ratio", "change", "base"))
}
\description{
Define outcome based on longitudinal outcomes
}
