% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baSaAlphaBeta}
\alias{baSaAlphaBeta}
\title{Get actuarial type I and II error in Sargent method}
\usage{
baSaAlphaBeta(bsizes, r, p0, p1, rho0, rho1)
}
\description{
Get actuarial type I and II error in Sargent method
}
