% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baSaDesign}
\alias{baSaDesign}
\title{Get Design type I and II error using Sargent method}
\usage{
baSaDesign(nmin, bsize, alpha, beta, p0, p1, rho0, rho1)
}
\description{
Get Design type I and II error using Sargent method
}
