% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baSimuBetaBin}
\alias{baSimuBetaBin}
\title{Simulate batch data from Beta-Binomial}
\usage{
baSimuBetaBin(bsize, p, rho = 0, seed = NULL, nreps = 1e+06)
}
\arguments{
\item{bsize}{batch size}

\item{p}{mean of the Beta distribution a/(a+b)}

\item{rho}{inter-class correlation 1/(a+b+1)}

\item{seed}{Seed}

\item{nrep}{number of repetitions}
}
\value{
Simulated binary outcomes by batches
}
\description{
Simulate batch data from Beta-Binomial
}
