% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baSimuError}
\alias{baSimuError}
\title{Simulate random error}
\usage{
baSimuError(
  n,
  error.type = c("normal", "skewed"),
  sig = 1,
  mu = 0,
  skew.mu = NULL,
  skew.phi = NULL,
  skew.noise = 0.001
)
}
\description{
Simulate random error
}
\details{
Skewed distribution: Prameterization: mu, phi; RNBINOM uses (n,p)
    with: phi = n, mu = n(1-p)/p; Mean: mu = n(1-p)/p #; Variances:
    mu(1+mu/phi) = n (1-p)/p^2
}
