% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baSimuTcell}
\alias{baSimuTcell}
\title{Simulate t-cell counts using Negative binomial}
\usage{
baSimuTcell(bsize, par.err, par.other, nreps = 1e+06, ...)
}
\description{
Simulate t-cell counts using Negative binomial
}
