% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cava_shiny.R
\name{cvShiny}
\alias{cvShiny}
\title{Run Web-Based \code{cava} application}
\usage{
cvShiny()
}
\description{
Call Shiny to run \code{cava} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI of \code{\link{visit}}.
}
\examples{
if(interactive()){
cvShiny()}

}
