% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cava_package.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example dataset}
\format{
A dataframe with the following variables:
\describe{
  \item{BatchID}{Batch ID}
  \item{Week}{Week when the experiment was done, i.e., time effect}
  \item{Concentration}{Concentration level}
  \item{Y}{Log transformed outcome}
}
}
\usage{
data(ex_data)
}
\description{
An example dataset from an assay experiment for eliciting ICC
}
