% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{expand_wildcards}
\alias{expand_wildcards}
\title{Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions}
\usage{
expand_wildcards(dictionary, vectors, type = c("glob", "regex"))
}
\arguments{
\item{dictionary}{a character vector of words containing wildcards}

\item{vectors}{A vectors object, e.g. as returned by load_fasttext}
}
\value{
a character vector containing the expanded terms
}
\description{
Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions
}
