% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{find_expansion_threshold}
\alias{find_expansion_threshold}
\title{Use word embedding to automatically expand seed set and compute metrics}
\usage{
find_expansion_threshold(dictionary, vectors, split = 0.5, k = 1000)
}
\arguments{
\item{dictionary}{a character vector of words containing wildcards}

\item{vectors}{A vectors object, e.g. as returned by load_fasttext}

\item{split}{Faction to use as training data}

\item{k}{Number of candidates to investigate}

\item{ft_model}{a FastTextR model}

\item{words}{a character vector of words}

\item{vocabulary}{If given, limit results to words from this vocabulary
(e.g. only words occurring in the target corpus)}
}
\value{
A tibble containing terms and (pseudo-)metrics
}
\description{
This will split the terms in train and test sets, use `nearest_neighbours`
to generate similar words, and test whether the generated words
occur in the test set, giving a pseudo-precision and pseudo-recall.
Note that both will be lower than actualy precision and recall as we assume
the seed words are incomplete (why else expand it), but it can serve as an
indication of how many terms to consider.
}
