% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{get_centroid}
\alias{get_centroid}
\title{Compute the normalized centroid vector for the given dictionary}
\usage{
get_centroid(dictionary, vectors)
}
\description{
Compute the normalized centroid vector for the given dictionary
}
