% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_fasttext}
\alias{load_fasttext}
\title{Load similarity information from a fasttext model}
\usage{
load_fasttext(filename, corpus = null, n = 10000)
}
\arguments{
\item{filename}{a filename pointing to a fasttext .bin model}

\item{corpus}{an optional corpus to use for word frequency information.
if given, should be a quanteda corpus, dfm or data frame with word and frequency as first two columns}
}
\value{
a list with a nxn similarity matrix and a frequency data frame
}
\description{
Load similarity information from a fasttext model
}
