% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{pairwise_similarities}
\alias{pairwise_similarities}
\title{Compute the pairwise similarities between all words in a dictionary}
\usage{
pairwise_similarities(dictionary, vectors)
}
\arguments{
\item{dictionary}{A character vector of words}

\item{vectors}{A vectors object, e.g. as returned by load_fasttext}
}
\value{
a (tibble) data frame with columns word1, word2, sim, frequency1 and frequency2
}
\description{
Compute the pairwise similarities between all words in a dictionary
}
