% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_present.R
\name{plot.ClsBaErr}
\alias{plot.ClsBaErr}
\title{S3 method for plotting the random error}
\usage{
\method{plot}{ClsBaErr}(x, opts = NULL, pos.legend = c(0.8, 0.6))
}
\description{
S3 method for plotting the random error
}
