% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{similar_words}
\alias{similar_words}
\title{Compute most similar words to a dictionary of seed words}
\usage{
similar_words(dictionary, vectors, antonyms = NULL)
}
\arguments{
\item{dictionary}{a character vector of words containing wildcards}

\item{vectors}{A vectors object, e.g. as returned by load_fasttext}

\item{k}{the (maximum) number of neighbours to return}
}
\value{
along-format tibble containing columns word and dist
}
\description{
Compute most similar words to a dictionary of seed words
}
