% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{similarity_graph}
\alias{similarity_graph}
\title{Create a similarity graph}
\usage{
similarity_graph(similarities, threshold = NULL, max_edges = NULL)
}
\arguments{
\item{similarities}{a data frame of word-word similarities (e.g. output from pairwise_similarity)}

\item{threshold}{the threshold for including a relation in the graph}
}
\value{
an igraph graph object
}
\description{
Create a similarity graph
}
