% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{similarity_to_centroid}
\alias{similarity_to_centroid}
\title{Compute the similarity of each word in the dictionary to the centroid}
\usage{
similarity_to_centroid(dictionary, vectors)
}
\arguments{
\item{dictionary}{a character vector of words}

\item{vectors}{A vectors object, e.g. as returned by load_fasttext}
}
\value{
a long-format tibble containing columns word and dist
}
\description{
Can be useful to find words that don't seem to belong in the dictionary
}
