% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bacSimonDesign}
\alias{bacSimonDesign}
\title{Simon's two-stage design}
\usage{
bacSimonDesign(y0, y1, nmax, nmin, bsize, alpha, beta)
}
\description{
@param y0 response 0
 @param y1 response 1
 @param nmax maximum size
 @param nmin minimum size
 @param bsize batch size
 @param alpha observed alpha from simulations
 @param beta observed beta from simulations.
}
