% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_design.R
\name{baDesignBetaBin}
\alias{baDesignBetaBin}
\title{Get designs for BetaBinomial designs}
\usage{
baDesignBetaBin(par.design, rho = 0, bsizes = 1, nmin = 10,
  nmax = 100, ...)
}
\description{
Get designs for BetaBinomial designs
}
