% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baGetBvrCvIcc}
\alias{baGetBvrCvIcc}
\title{Get CV, batch effect variance ratio and ICC}
\usage{
baGetBvrCvIcc(par.err, par.other, rates, f.simu = baSimuTcell, ...)
}
\arguments{
\item{par.err}{a list of gamma, delta, epsilon error}

\item{par.other}{other parameters including u0, u1, v, and beta}

\item{rates}{response rates}

\item{f.simu}{simulating function}
}
\value{
a list containing CV, batch effect variance ratio, and ICC
}
\description{
Get CV, batch effect variance ratio and ICC
}
