% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baGetCuts}
\alias{baGetCuts}
\title{Get cut off of the outcome to get given response rates}
\usage{
baGetCuts(par.err, par.other, rates, f.simu = baSimuTcell, ..., seed = NULL)
}
\description{
Get cut off of the outcome to get given response rates
}
