% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_tools.R
\name{baGetOutcome}
\alias{baGetOutcome}
\title{Define outcome based on longitudinal outcomes}
\usage{
baGetOutcome(mat.y, type = c("ratio", "change", "base"))
}
\arguments{
\item{mat.y}{outcomes represented as a matrix}

\item{type}{type of longitudinal outcomes}
}
\value{
if type is "base", the first column is returned. Otherwise
    the ratio of last column and first column is returned.
}
\description{
Define outcome based on longitudinal outcomes
}
