% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_design.R
\name{baGetRejRate}
\alias{baGetRejRate}
\title{Get type I error and power given y0, y1, n1, r1, n, r}
\usage{
baGetRejRate(yp0, yp1, n1, r1, nt, r, bsize = 1)
}
\description{
Get type I error and power given y0, y1, n1, r1, n, r
}
