% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_present.R
\name{baPltAllTcell}
\alias{baPltAllTcell}
\title{Plot histogram of t-cells Y0 Y1 and Y1/Y0 for all scenarios}
\usage{
baPltAllTcell(
  lst.par,
  xlims = list(c(-1, 1), c(0, 2000), c(0, 20)),
  ylims = c(9, 0.2, 0.3),
  label = "Scenario",
  ...
)
}
\arguments{
\item{lst.par}{a list containing par.err and par.other as described in \code{baPltTcell}}

\item{xlims}{x limit of the plot}

\item{ylims}{y limit of the plot}

\item{label}{label for the plot}
}
\description{
Plot histogram of t-cells Y0 Y1 and Y1/Y0 for all scenarios
}
