% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_present.R
\name{baPltTcell}
\alias{baPltTcell}
\title{Plot histogram of t-cells Y0 Y1 and Y1/Y0 for a given scenario}
\usage{
baPltTcell(
  par.err,
  par.other,
  nreps = 1e+05,
  pos.legend = c(0.8, 0.6),
  f.simu = baSimuTcell,
  fname = NULL,
  ...,
  ry.quants = c(0.1, 0.5, 0.9)
)
}
\arguments{
\item{par.err}{a list of gamma, delta, epsilon error}

\item{par.other}{other parameters including u0, u1, v, and beta}

\item{nreps}{number of repetitions}

\item{pos.legend}{legend position}

\item{f.simu}{simulating function}

\item{fname}{file name to save to}

\item{...}{reserved parameters}

\item{ry.quants}{quantiles}
}
\value{
a plot of histogram of t-cells Y0 Y1 and Y1/Y0 for a given scenario
}
\description{
Plot histogram of t-cells Y0 Y1 and Y1/Y0 for a given scenario
}
