% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baSaGetVar}
\alias{baSaGetVar}
\title{Get variance given batch sizes, response rate and ICC
 @export}
\usage{
baSaGetVar(bsizes, p, rho)
}
\description{
Get variance given batch sizes, response rate and ICC
 @export
}
